requirements.txt
# ============ Core ============
torch>=2.1.0
transformers>=4.45.0
timm>=1.0.8

# ============ LangChain & Upstage ============
langchain-core>=0.3.0
langchain-community>=0.2.0
langchain-upstage>=0.1.0
langchain-huggingface>=0.0.2
sentence-transformers>=2.2.2
upstage==0.1
openai>=1.48.0

# ============ DB & Vector Search ============
pgvector>=0.2.4
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.30
faiss-cpu>=1.8.0
langchain-postgres>=0.0.9

# ============ PDF & OCR ============
pdfminer.six>=20221105
pdf2image>=1.17.0
pillow>=10.0.0

# ============ Server & Utils ============
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0




